Renamed AlertRoot to Alert
